requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
watchdog==3.0.0
psutil==5.9.6
requests==2.31.0